refactor: Generate claims in the same way

+++ token/jwt/claims_id_token_test.go
@@ -72,7 +72,7 @@ func TestIDTokenClaimsToMap(t *testing.T) {
 		"baz":       idTokenClaims.Extra["baz"],
 		"at_hash":   idTokenClaims.AccessTokenHash,
 		"c_hash":    idTokenClaims.CodeHash,
-		"auth_time": idTokenClaims.AuthTime.Unix(),
+		"auth_time": float64(idTokenClaims.AuthTime.Unix()),
 		"acr":       idTokenClaims.AuthenticationContextClassReference,
 		"amr":       idTokenClaims.AuthenticationMethodsReference,
 	}, idTokenClaims.ToMap())
